% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_active.R
\name{active_new}
\alias{active_new}
\title{Active algorithm constructor.}
\usage{
active_new(
  pipeline = NULL,
  meta = NULL,
  names = NULL,
  queue = NULL,
  reporter = NULL,
  garbage_collection = NULL
)
}
\arguments{
\item{pipeline}{Pipeline object.}

\item{meta}{Meta object.}

\item{names}{Character, names of targets.}

\item{queue}{Character, type of queue.}

\item{reporter}{Character, type of reporter.}

\item{garbage_collection}{Logical, whether to
periodically run garbage collection.}
}
\description{
Not a user-side function.
Used to build external HPC scheduling algorithms.
}
\keyword{internal}
