% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_algorithm.R
\name{algorithm_class}
\alias{algorithm_class}
\alias{tar_algorithm}
\title{Abstract class for algorithm objects.}
\description{
Not a user-side R6 class.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pipeline}}{Pipeline object.}

\item{\code{meta}}{Meta object.}

\item{\code{scheduler}}{Scheduler object.}

\item{\code{names}}{Character, names of targets.}

\item{\code{queue}}{Character, name of the queue type.}

\item{\code{reporter}}{Character, name of the reporter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{algorithm_class$new()}}
\item \href{#method-update_scheduler}{\code{algorithm_class$update_scheduler()}}
\item \href{#method-validate}{\code{algorithm_class$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an active algorithm object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{algorithm_class$new(
  pipeline = NULL,
  meta = NULL,
  names = NULL,
  queue = NULL,
  reporter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pipeline}}{Pipeline object.}

\item{\code{meta}}{Meta object.}

\item{\code{names}}{Character, names of targets.}

\item{\code{queue}}{Character, type of queue.}

\item{\code{reporter}}{Character, type of reporter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_scheduler"></a>}}
\if{latex}{\out{\hypertarget{method-update_scheduler}{}}}
\subsection{Method \code{update_scheduler()}}{
Create the scheduler from the
pipeline and settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{algorithm_class$update_scheduler()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{algorithm_class$validate()}\if{html}{\out{</div>}}
}

}
}
