% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{assert_callr_function}
\alias{assert_callr_function}
\title{Internal infrastructure function.}
\usage{
assert_callr_function(callr_function)
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
