% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{assert_list}
\alias{assert_list}
\title{Internal infrastructure function.}
\usage{
assert_list(x, msg = NULL)
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
