% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{assert_script}
\alias{assert_script}
\title{Assert the target script.}
\usage{
assert_script(script)
}
\arguments{
\item{script}{Character of length 1, path to the target script file.}
}
\value{
\code{NULL} (invisibly) if a valid target script exists.
Otherwise, an error of class \code{"tar_condition_validate"}
is thrown.
}
\description{
Exported for internal purposes only. Not a
user-side function. Do not invoke directly.
}
\keyword{internal}
