% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{assert_target_script}
\alias{assert_target_script}
\title{Internal infrastructure function.}
\usage{
assert_target_script()
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
