% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_callr.R
\name{callr_args_default}
\alias{callr_args_default}
\title{Deprecated: \code{callr} arguments.}
\usage{
callr_args_default(callr_function, reporter = NULL)
}
\arguments{
\item{callr_function}{A function from the \code{callr} package
that starts an external R process.}

\item{reporter}{Character of length 1, choice of reporter
for \code{\link[=tar_make]{tar_make()}} or a related function.}
}
\value{
A list of arguments to \code{callr_function}.
}
\description{
Deprecated on 2022-08-05 (\code{targets} version 0.13.1).
Please use \code{\link[=tar_callr_args_default]{tar_callr_args_default()}} instead.
}
\details{
Not a user-side function. Do not invoke directly.
Exported for internal purposes only.
}
\examples{
tar_callr_args_default(callr::r)
}
\keyword{internal}
