% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_callr.R
\name{callr_outer}
\alias{callr_outer}
\title{Internal infrastructure function.}
\usage{
callr_outer(
  targets_function,
  targets_arguments,
  callr_function,
  callr_arguments
)
}
\arguments{
\item{targets_function}{A \code{targets} function.}

\item{targets_arguments}{List of arguments to \code{targets_function}.}

\item{callr_arguments}{List of arguments to  \code{callr_function}.}

\item{callr_Function}{A \code{callr} function.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
