% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_dynamic_map.R
\name{dynamic-map}
\alias{dynamic-map}
\alias{map}
\title{Dynamic map}
\description{
Iterate over one or more targets in sequence.
}
\details{
\code{pattern = map(x, y)} creates one dynamic branch for
the first element of \code{x} and the first element of \code{y},
another for the second element of \code{x} and the second element of \code{y},
and so on.
}
\examples{
# _targets.R
tar_pipeline(
  tar_target(x, seq_len(2)),
  tar_target(y, head(letters, 2)),
  tar_target(dynamic, c(x, y), pattern = map(x, y)) # 2 branches
)
}
\seealso{
Other Dynamic branching: 
\code{\link{dynamic-cross}}
}
\concept{Dynamic branching}
