% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{if_any}
\alias{if_any}
\title{Non-vectorized if/else.}
\usage{
if_any(condition, x, y)
}
\arguments{
\item{condition}{A value or expression that evaluates to
a logical of length 1 when invoked through \code{any()}.}

\item{`x`}{Value to return if \code{any(condition)}
evaluates to \code{TRUE}.}

\item{`y`}{Value to return if \code{any(condition)}
does not evaluate to \code{TRUE}.}
}
\value{
\code{x} if \code{any(condition)} evaluates to \code{TRUE},
otherwise \code{y}.
}
\description{
Internal to \code{targets}. Not a user-side function.
Do not invoke directly. Only exported for the purpose
of using \code{callr} functions.
}
\examples{
if_any(TRUE, 1, 2)
if_any(FALSE, 1, 2)
}
\keyword{internal}
