% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_local.R
\name{local_class}
\alias{local_class}
\alias{tar_local}
\title{Abstract class for local algorithm objects.}
\description{
Not a user-side R6 class.
}
\keyword{internal}
\section{Super classes}{
\code{\link[targets:tar_algorithm]{targets::tar_algorithm}} -> \code{\link[targets:tar_active]{targets::tar_active}} -> \code{tar_local}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-assert_deployment}{\code{local_class$assert_deployment()}}
\item \href{#method-run_target}{\code{local_class$run_target()}}
\item \href{#method-process_target}{\code{local_class$process_target()}}
\item \href{#method-process_next}{\code{local_class$process_next()}}
\item \href{#method-run}{\code{local_class$run()}}
\item \href{#method-validate}{\code{local_class$validate()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_algorithm" data-id="update_scheduler">}\href{../../targets/html/tar_algorithm.html#method-update_scheduler}{\code{targets::tar_algorithm$update_scheduler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="end">}\href{../../targets/html/tar_active.html#method-end}{\code{targets::tar_active$end()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="ensure_meta">}\href{../../targets/html/tar_active.html#method-ensure_meta}{\code{targets::tar_active$ensure_meta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="initialize">}\href{../../targets/html/tar_active.html#method-initialize}{\code{targets::tar_active$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="run_gc">}\href{../../targets/html/tar_active.html#method-run_gc}{\code{targets::tar_active$run_gc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="start">}\href{../../targets/html/tar_active.html#method-start}{\code{targets::tar_active$start()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="unload_transient">}\href{../../targets/html/tar_active.html#method-unload_transient}{\code{targets::tar_active$unload_transient()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="unserialize_target">}\href{../../targets/html/tar_active.html#method-unserialize_target}{\code{targets::tar_active$unserialize_target()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-assert_deployment}{}}}
\subsection{Method \code{assert_deployment()}}{
Only applies to prelocal objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_class$assert_deployment(target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Target object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_target"></a>}}
\if{latex}{\out{\hypertarget{method-run_target}{}}}
\subsection{Method \code{run_target()}}{
Run a target.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_class$run_target(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a target to run.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process_target"></a>}}
\if{latex}{\out{\hypertarget{method-process_target}{}}}
\subsection{Method \code{process_target()}}{
Process a target.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_class$process_target(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a target to process.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process_next"></a>}}
\if{latex}{\out{\hypertarget{method-process_next}{}}}
\subsection{Method \code{process_next()}}{
Process the next target.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_class$process_next()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run the algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_class$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_class$validate()}\if{html}{\out{</div>}}
}

}
}
