% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_memory.R
\name{memory_clear_objects}
\alias{memory_clear_objects}
\title{Internal infrastructure function.}
\usage{
memory_clear_objects(memory)
}
\arguments{
\item{memory}{Memory object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
