% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_memory.R
\name{memory_del_objects}
\alias{memory_del_objects}
\title{Internal infrastructure function.}
\usage{
memory_del_objects(memory, names)
}
\arguments{
\item{memory}{Memory object.}

\item{names}{Names of the data objects to delete.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
