% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_memory.R
\name{memory_exists_object}
\alias{memory_exists_object}
\title{Internal infrastructure function.}
\usage{
memory_exists_object(memory, name)
}
\arguments{
\item{memory}{Memory object.}

\item{name}{Name of the data object to check.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
