% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_memory.R
\name{memory_init}
\alias{memory_init}
\title{Internal infrastructure function.}
\usage{
memory_init(envir = new.env(parent = emptyenv()))
}
\arguments{
\item{envir}{Environment for the data.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
