% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_memory.R
\name{memory_set_object}
\alias{memory_set_object}
\title{Internal infrastructure function.}
\usage{
memory_set_object(memory, name, object)
}
\arguments{
\item{memory}{Memory object.}

\item{name}{Name of the data object to set.}

\item{object}{Data object to set.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
