% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_meta.R
\name{meta_init}
\alias{meta_init}
\title{Internal infrastructure function.}
\usage{
meta_init()
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
