% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{path_objects_dir_cloud}
\alias{path_objects_dir_cloud}
\title{Default pseudo-directory path of target data in the cloud}
\usage{
path_objects_dir_cloud()
}
\value{
Character of length,
default pseudo-directory path of target data in the cloud.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
path_objects_dir_cloud()
}
\keyword{internal}
