% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{path_script_r}
\alias{path_script_r}
\title{Target script helper directory path.}
\usage{
path_script_r(path_script)
}
\arguments{
\item{path_script}{Character of length 1, path to the
target script file (default: \verb{_targets.R}).}
}
\value{
Character of length 1, target script
helper directory path.
}
\description{
Internal function for Target Markdown.
Not a user-side function. Do not invoke directly.
}
\examples{
path_script_r()
}
\keyword{internal}
