% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pipeline.R
\name{pipeline_get_names}
\alias{pipeline_get_names}
\title{Internal infrastructure function.}
\usage{
pipeline_get_names(pipeline)
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
