% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pipeline.R
\name{pipeline_get_target}
\alias{pipeline_get_target}
\title{Internal infrastructure function.}
\usage{
pipeline_get_target(pipeline, name)
}
\arguments{
\item{pipeline}{Pipeline object.}

\item{name}{Character, name of the target to get.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
