% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pipeline.R
\name{pipeline_set_target}
\alias{pipeline_set_target}
\title{Internal infrastructure function.}
\usage{
pipeline_set_target(pipeline, target)
}
\arguments{
\item{pipeline}{Pipeline object.}

\item{target}{Target object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
