% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pipeline.R
\name{pipeline_validate_lite}
\alias{pipeline_validate_lite}
\title{Internal infrastructure function.}
\usage{
pipeline_validate_lite(pipeline)
}
\arguments{
\item{pipeline}{Pipeline object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
