% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_prelocal.R
\name{prelocal_class}
\alias{prelocal_class}
\alias{tar_prelocal}
\title{Abstract class for prelocal algorithm objects.}
\description{
Not a user-side R6 class.
Used to build external HPC scheduling algorithms.
}
\keyword{internal}
\section{Super classes}{
\code{\link[targets:tar_algorithm]{targets::tar_algorithm}} -> \code{\link[targets:tar_active]{targets::tar_active}} -> \code{\link[targets:tar_local]{targets::tar_local}} -> \code{tar_prelocal}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{prelocal_class$new()}}
\item \href{#method-start}{\code{prelocal_class$start()}}
\item \href{#method-end}{\code{prelocal_class$end()}}
\item \href{#method-assert_deployment}{\code{prelocal_class$assert_deployment()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_algorithm" data-id="update_scheduler">}\href{../../targets/html/tar_algorithm.html#method-update_scheduler}{\code{targets::tar_algorithm$update_scheduler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="ensure_meta">}\href{../../targets/html/tar_active.html#method-ensure_meta}{\code{targets::tar_active$ensure_meta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="run_gc">}\href{../../targets/html/tar_active.html#method-run_gc}{\code{targets::tar_active$run_gc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="unload_transient">}\href{../../targets/html/tar_active.html#method-unload_transient}{\code{targets::tar_active$unload_transient()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_active" data-id="unserialize_target">}\href{../../targets/html/tar_active.html#method-unserialize_target}{\code{targets::tar_active$unserialize_target()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_local" data-id="process_next">}\href{../../targets/html/tar_local.html#method-process_next}{\code{targets::tar_local$process_next()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_local" data-id="process_target">}\href{../../targets/html/tar_local.html#method-process_target}{\code{targets::tar_local$process_target()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_local" data-id="run">}\href{../../targets/html/tar_local.html#method-run}{\code{targets::tar_local$run()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_local" data-id="run_target">}\href{../../targets/html/tar_local.html#method-run_target}{\code{targets::tar_local$run_target()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="targets" data-topic="tar_local" data-id="validate">}\href{../../targets/html/tar_local.html#method-validate}{\code{targets::tar_local$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a prelocal algorithm object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prelocal_class$new(
  pipeline = NULL,
  meta = NULL,
  names = NULL,
  queue = NULL,
  reporter = NULL,
  garbage_collection = NULL,
  scheduler = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pipeline}}{Pipeline object.}

\item{\code{meta}}{Meta object.}

\item{\code{names}}{Character, names of targets.}

\item{\code{queue}}{Character, type of queue.}

\item{\code{reporter}}{Character, type of reporter.}

\item{\code{garbage_collection}}{Logical, whether to
periodically run garbage collection.}

\item{\code{scheduler}}{Scheduler object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Start the algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prelocal_class$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end"></a>}}
\if{latex}{\out{\hypertarget{method-end}{}}}
\subsection{Method \code{end()}}{
End the algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prelocal_class$end()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-assert_deployment}{}}}
\subsection{Method \code{assert_deployment()}}{
Assert that the target runs locally
and exit otherwise. The special error condition defers
control to the calling algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prelocal_class$assert_deployment(target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Target object.}
}
\if{html}{\out{</div>}}
}
}
}
