% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{restore_config}
\alias{restore_config}
\title{Restore config settings (after a function call).}
\usage{
restore_config(old_config)
}
\arguments{
\item{script}{Character of length 1, path to the target script file.}

\item{store}{Character of length 1, path to the data store.}
}
\value{
\code{NULL} (invisibly) if a valid target script exists.
Otherwise, an error of class \code{"tar_condition_validate"}
is thrown.
}
\description{
Exported for internal purposes only. Not a
user-side function. Do not invoke directly.
}
\keyword{internal}
