% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin_tar_glimpse.R
\name{rstudio_addin_tar_glimpse}
\alias{rstudio_addin_tar_glimpse}
\title{RStudio addin to call \code{\link[=tar_glimpse]{tar_glimpse()}}.}
\usage{
rstudio_addin_tar_glimpse()
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
