% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin_tar_make_bg.R
\name{rstudio_addin_tar_make_bg}
\alias{rstudio_addin_tar_make_bg}
\title{RStudio addin to run \code{\link[=tar_make]{tar_make()}} in the background.}
\usage{
rstudio_addin_tar_make_bg()
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
