% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin_tar_outdated.R
\name{rstudio_addin_tar_outdated}
\alias{rstudio_addin_tar_outdated}
\title{RStudio addin to call \code{\link[=tar_outdated]{tar_outdated()}}.}
\usage{
rstudio_addin_tar_outdated()
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
