% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin_tar_read.R
\name{rstudio_addin_tar_read}
\alias{rstudio_addin_tar_read}
\title{RStudio addin to call \code{\link[=tar_read]{tar_read()}} on the symbol at the cursor.}
\usage{
rstudio_addin_tar_read(context = NULL)
}
\arguments{
\item{context}{RStudio API context from
\code{rstudioapi::getActiveDocumentContext()}.}
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
