% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin_tar_target.R
\name{rstudio_addin_tar_target}
\alias{rstudio_addin_tar_target}
\title{RStudio addin to insert \code{"tar_target()"} at the cursor.}
\usage{
rstudio_addin_tar_target(context = NULL)
}
\arguments{
\item{context}{RStudio API context from
\code{rstudioapi::getActiveDocumentContext()}.}
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
