% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_store_custom.R
\name{store_assert_format_setting.store_custom}
\alias{store_assert_format_setting.store_custom}
\title{#' @export}
\usage{
\method{store_assert_format_setting}{store_custom}(class)
}
\description{
#' @export
}
