% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pipeline.R
\name{tar_as_pipeline}
\alias{tar_as_pipeline}
\title{Convert to a pipeline object.}
\usage{
tar_as_pipeline(x)
}
\arguments{
\item{x}{A list of target objects or a pipeline object.}
}
\value{
An object of class \code{"tar_pipeline"}.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
