% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_bind.R
\name{tar_bind}
\alias{tar_bind}
\title{Combine pipelines.}
\usage{
tar_bind(...)
}
\arguments{
\item{...}{Pipeline objects or nested lists of pipeline objects.
You can generate a pipeline object with \code{\link[=tar_pipeline]{tar_pipeline()}}.}
}
\description{
Combine multiple pipeline objects
into a single pipeline object.
}
\examples{
# Include in _targets.R:
pipeline1 <- tar_pipeline(tar_target(x, 1))
pipeline2 <- tar_pipeline(tar_target(y, 2))
# Pipeline object with both targets:
pipeline <- tar_bind(pipeline1, pipeline2)
}
