% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_repository_cas_local.R
\name{tar_cas_d}
\alias{tar_cas_d}
\title{Local CAS download.}
\usage{
tar_cas_d(cas, key, path)
}
\arguments{
\item{cas}{File path to the CAS repository. \code{NULL} to default to
\code{file.path(tar_config_get("store"), "cas")} (usually \code{"_targets/cas/"}).}

\item{key}{Key of the object in the CAS system.}

\item{path}{Staging path of the file.}
}
\value{
Called for its side effects.
}
\description{
For internal use only.
}
\keyword{internal}
