% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_repository_cas_local.R
\name{tar_cas_e}
\alias{tar_cas_e}
\title{Check existence in local CAS.}
\usage{
tar_cas_e(cas, key)
}
\arguments{
\item{cas}{File path to the CAS repository. \code{NULL} to default to
\code{file.path(tar_config_get("store"), "cas")} (usually \code{"_targets/cas/"}).}

\item{key}{Key of the object in the CAS system.}
}
\value{
Character vector of keys (metadata hashes) found in the CAS system.
}
\description{
For internal use only.
}
\details{
The short function name helps reduce the size of the
\code{\link[=tar_repository_cas]{tar_repository_cas()}} format string and save space in the metadata.
}
\keyword{internal}
