% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_repository_cas_local.R
\name{tar_cas_l}
\alias{tar_cas_l}
\title{List keys in local CAS.}
\usage{
tar_cas_l(cas, keys)
}
\arguments{
\item{cas}{File path to the CAS repository. \code{NULL} to default to
\code{file.path(tar_config_get("store"), "cas")} (usually \code{"_targets/cas/"}).}

\item{keys}{Character vector of keys in the metadata hashes
(\code{tar_meta()$data}). Used to restrict the output of the return value
to avoid listing all the potentially millions of files in the
CAS system.}
}
\value{
Character vector of keys (metadata hashes) found in the CAS system.
}
\description{
For internal use only.
}
\details{
The short function name helps reduce the size of the
\code{\link[=tar_repository_cas]{tar_repository_cas()}} format string and save space in the metadata.
}
\keyword{internal}
