% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_repository_cas_local.R
\name{tar_cas_u}
\alias{tar_cas_u}
\title{Local CAS upload.}
\usage{
tar_cas_u(cas, key, path)
}
\arguments{
\item{cas}{File path to the CAS repository. \code{NULL} to default to
\code{file.path(tar_config_get("store"), "cas")} (usually \code{"_targets/cas/"}).}

\item{key}{Key of the object in the CAS system.}

\item{path}{Staging path of the file.}
}
\value{
Called for its side effects.
}
\description{
For internal use only.
}
\details{
The short function name helps reduce the size of the
\code{\link[=tar_repository_cas]{tar_repository_cas()}} format string and save space in the metadata.
}
\keyword{internal}
