% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_condition.R
\name{tar_condition}
\alias{tar_condition}
\alias{tar_message_run}
\alias{tar_throw_file}
\alias{tar_throw_run}
\alias{tar_throw_validate}
\alias{tar_warn_deprecate}
\alias{tar_warn_run}
\alias{tar_warn_validate}
\alias{tar_error}
\alias{tar_warning}
\alias{tar_message}
\title{Conditions}
\usage{
tar_message_run(...)

tar_throw_file(...)

tar_throw_run(...)

tar_throw_validate(...)

tar_warn_deprecate(...)

tar_warn_run(...)

tar_warn_validate(...)

tar_error(message, class)

tar_warning(message, class)

tar_message(message, class)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{message}{Character of length 1, text of the message.}

\item{class}{Character vector of S3 classes of the message.}
}
\description{
These functions throw custom
\code{targets}-specific error conditions.
Useful for error handling in packages built on top of \code{targets}.
}
\examples{
try(tar_throw_validate("something is not valid"))
}
\seealso{
Other utilities to extend targets: 
\code{\link{tar_assert}},
\code{\link{tar_dir}()},
\code{\link{tar_language}},
\code{\link{tar_test}()}
}
\concept{utilities to extend targets}
