% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_condition.R
\name{tar_condition_traced}
\alias{tar_condition_traced}
\title{Contain an error condition and formatted traceback.}
\usage{
tar_condition_traced(condition, trace)
}
\arguments{
\item{condition}{An error condition object thrown by \code{stop()} or
\code{rlang::abort()}.}

\item{trace}{A raw return value from \code{.traceback()}.}
}
\value{
Contain an error condition and formatted traceback.
}
\description{
Not a user-side function.
}
\keyword{internal}
