% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_counter.R
\name{tar_counter}
\alias{tar_counter}
\title{Create a counter object.}
\usage{
tar_counter(names = NULL)
}
\arguments{
\item{names}{Character vector of names to set in the counter.}
}
\description{
Internal function. Not for users.
}
\examples{
tar_counter(names = "x")
}
\keyword{internal}
