% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_deps.R
\name{tar_deps_raw}
\alias{tar_deps_raw}
\title{Code dependencies (raw version)}
\usage{
tar_deps_raw(expr)
}
\arguments{
\item{expr}{An R expression object or function.}
}
\value{
Character vector of the dependencies of a function or expression.
}
\description{
Same as \code{\link[=tar_deps]{tar_deps()}} except \code{expr} must already be an
unquoted function or expression object.
}
\examples{
tar_deps_raw(quote(x <- y + z))
tar_deps_raw(
  quote({
    x <- 1
    x + a
  })
)
tar_deps_raw(function(a = b) map_dfr(data, ~do_row(.x)))
}
