% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_destroy.R
\name{tar_destroy}
\alias{tar_destroy}
\title{Destroy a section or all of the \verb{_targets/} data store in the
current working directory}
\usage{
tar_destroy(what = c("all", "meta", "progress", "objects"))
}
\arguments{
\item{what}{Character of length 1, what to destroy. Choices:
\itemize{
\item \code{"all"}: destroy the entire data store.
\item \code{"meta"}: just delete the metadata file at \verb{_targets/meta/meta},
which invalidates all the targets but keeps the data.
\item \code{"progress"}: just delete the progress data file at
\verb{_targets/meta/progress}, which resets the progress tracking info.
\item \code{"objects"}: delete all the target return values in \verb{_targets/objects/}
but keep progress and metadata. Dynamic files are not deleted this way.
}}
}
\description{
Destroy \verb{_targets/} data store in the current working
directory. Optionally, just destroy part of the data store.
}
\examples{
\dontrun{
tar_dir({
tar_script(tar_pipeline(tar_target(x, 1 + 1)))
tar_make() # Creates the _targets/ data store.
tar_destroy()
print(file.exists("_targets")) # Should be FALSE.
})
}
}
