% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_dynamic_sample.R
\name{tar_dynamic_sample}
\alias{tar_dynamic_sample}
\alias{sample}
\title{Dynamic sample}
\usage{
tar_dynamic_sample(x, n = 1L, seed = 0L)
}
\arguments{
\item{x}{Data frame with columns to represent targets and
elements to represent branches or buds. Column names must
be unique.}

\item{n}{Number of branches to create. Cannot exceed \code{nrow(x)}.}

\item{seed}{Integer of length 1, random number generator seed.
Within a target's pattern, e.g. \code{pattern = sample(x, n = 5)}
in \code{\link[=tar_target]{tar_target()}}, there is not an option to supply a seed.
Instead, the seed is generated from the target name in
a deterministic manner for the sake of reproducibility.}
}
\description{
Branch over the first few elements of a pattern.
}
\details{
In \code{\link[=tar_target]{tar_target()}}, \code{pattern = sample(x, n = 5)} branches over
a random 5 elements of \code{x}. The number random seed is set to the seed
of the target so branch selection is reproducible.
\code{sample()} can be composed with other kinds of patterns: for example,
\code{pattern = sample(map(x, y), n = 5)}.
To test and experiment with this behavior outside
the pipeline, use the \code{tar_dynamic_sample()} function.
}
\examples{
# To use dynamic sample for real in a pipeline,
# call sample() in a target's pattern.
# The following code goes at the bottom of _targets.R.
tar_pipeline(
  tar_target(x, seq_len(26)),
  tar_target(dynamic, c(x, y), pattern = sample(x, n = 2)) # 2 branches
)
# To explore dynamic sample() outside a pipeline,
# use the tar_dynamic_sample() helper function.
x <- data.frame(x = seq_len(26))
tar_dynamic_sample(x, n = 2)
# You can experiment with more complicated patterns.
# The following is a representation of pattern = sample(map(x, y), n = 2).
y <- data.frame(y = letters)
tar_dynamic_sample(tar_dynamic_map(x, y), n = 2)
}
\seealso{
Other Dynamic branching: 
\code{\link{tar_dynamic_cross}()},
\code{\link{tar_dynamic_head}()},
\code{\link{tar_dynamic_map}()},
\code{\link{tar_dynamic_tail}()}
}
\concept{Dynamic branching}
