% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_edit.R
\name{tar_edit}
\alias{tar_edit}
\title{Open the \code{targets} script file.}
\usage{
tar_edit()
}
\description{
Open the \code{targets} script file for editing.
Requires the \code{usethis} package.
}
\details{
The \code{targets} script file is an R code file
that defines the pipeline. The default path is \verb{_targets.R},
but the default for the current project
can be configured with \code{\link[=tar_config_set]{tar_config_set()}}.
}
\seealso{
Other scripts: 
\code{\link{tar_github_actions}()},
\code{\link{tar_helper_raw}()},
\code{\link{tar_helper}()},
\code{\link{tar_renv}()},
\code{\link{tar_script}()}
}
\concept{scripts}
