% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_envir.R
\name{tar_envir}
\alias{tar_envir}
\title{Deprecated: get the environment of the current target.}
\usage{
tar_envir(default = parent.frame())
}
\arguments{
\item{default}{Environment, value to return if \code{tar_envir()}
is called on its own outside a \code{targets} pipeline.
Since \verb{targets >= 0.3.0}, the default is always the return value.}
}
\value{
An environment, just the parent frame.
In \code{targets} <= 0.2.0, \code{tar_envir()} returned the top-level
environment where the target was running. Recent
infrastructure changes make this no longer possible.
}
\description{
Deprecated in \code{targets} version 0.3.0 (2020-03-24).
Gets the parent frame. In earlier versions,
\code{tar_envir()} could return the environment where a
target runs its command (inherits from \code{tar_option_get("envir")}).
Recent infrastructure changes make this no longer possible.
}
\details{
Users should not call \code{tar_envir()} directly because accidental
modifications to \code{parent.env(tar_envir())} could break the pipeline.
\code{tar_envir()} only exists in order to support third-party interface
packages such as \code{tarchetypes}.
}
\examples{
tar_envir()
}
\keyword{internal}
