% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_exist_meta.R
\name{tar_exist_meta}
\alias{tar_exist_meta}
\title{Check if target metadata exists.}
\usage{
tar_exist_meta()
}
\value{
Logical of length 1, whether the current project's metadata exists.
}
\description{
Check if the target metadata file \verb{_targets/meta/meta}
exists for the current project.
}
\details{
To learn more about local storage in \code{targets}, visit
\url{https://books.ropensci.org/targets/files.html#internal-files}.
}
\examples{
tar_exist_meta()
}
