% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_exist_process.R
\name{tar_exist_process}
\alias{tar_exist_process}
\title{Check if process metadata exists.}
\usage{
tar_exist_process()
}
\value{
Logical of length 1, whether the current project's metadata exists.
}
\description{
Check if the process metadata file \verb{_targets/meta/process}
exists for the current project.
}
\details{
To learn more about local storage in \code{targets}, visit
\url{https://books.ropensci.org/targets/files.html#internal-files}.
}
\examples{
tar_exist_process()
}
