% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_exist_script.R
\name{tar_exist_script}
\alias{tar_exist_script}
\title{Check if the \code{targets} script file exists.}
\usage{
tar_exist_script()
}
\value{
Logical of length 1, whether the current project's metadata exists.
}
\description{
Check if the \code{targets} script file exists for the
current project. The \code{targets} script is \verb{_targets.R} by default,
but the path can be configured for the current project
using \code{\link[=tar_config_set]{tar_config_set()}}.
}
\examples{
tar_exist_script()
}
\seealso{
Other existence: 
\code{\link{tar_exist_meta}()},
\code{\link{tar_exist_objects}()},
\code{\link{tar_exist_process}()},
\code{\link{tar_exist_progress}()}
}
\concept{existence}
