% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_knitr_engine.R
\name{tar_knitr_engine}
\alias{tar_knitr_engine}
\title{Target Markdown \code{knitr} engine}
\usage{
tar_knitr_engine(options)
}
\arguments{
\item{options}{A named list of \code{knitr} chunk options.}
}
\value{
Character, output generated from \code{knitr::engine_output()}.
}
\description{
\code{knitr} language engine that runs \code{{targets}}
code chunks in Target Markdown.
}
\examples{
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
# Register the engine.
if (requireNamespace("knitr", quietly = TRUE)) {
  knitr::knit_engines$set(targets = targets::tar_knitr_engine)
}
# Then, {targets} code chunks in a knitr report will run
# as described at https://books.ropensci.org/targets/markdown.html.
}
}
\seealso{
\url{https://books.ropensci.org/targets/markdown.html}

Other targetopia: 
\code{\link{tar_dir}()},
\code{\link{tar_test}()}
}
\concept{targetopia}
