% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_make.R
\name{tar_make}
\alias{tar_make}
\title{Run a pipeline of targets.}
\usage{
tar_make(
  names = NULL,
  reporter = "verbose",
  garbage_collection = FALSE,
  callr_function = callr::r,
  callr_arguments = list()
)
}
\arguments{
\item{names}{Names of the targets to build or check. Set to \code{NULL} to
check/build all the targets (default). Otherwise, you can supply
symbols, a character vector, or \code{tidyselect} helpers like \code{\link[=starts_with]{starts_with()}}.}

\item{reporter}{Character of length 1, name of the reporter to user.
Controls how messages are printed as targets run in the pipeline. Choices:
\itemize{
\item \code{"verbose"}: print one message for each target that runs (default).
\item \code{"silent"}: print nothing.
\item \code{"timestamp"}: print a time-stamped message for each target that runs.
\item \code{"summary"}: print a running total of the number of each targets in
each status category (queued, running, skipped, build, cancelled,
or errored).
}}

\item{garbage_collection}{Logical, whether to run \code{base::gc()}
between targets. The pipeline will run slower but consume less memory.}

\item{callr_function}{A function from \code{callr} to start a fresh clean R
process to do the work. Set to \code{NULL} to do the work in the current R
session, which is not reproducible and only recommended for
debugging purposes.}

\item{callr_arguments}{A list of arguments to \code{callr_function}.}
}
\value{
Nothing.
}
\description{
Run the pipeline you defined in \verb{_targets.R}. \code{tar_make()}
runs the correct targets in the correct order and stores the return
values in \verb{_targets/objects/}.
}
\examples{
tar_dir({
tar_script({
  tar_options()
  tar_pipeline(tar_target(x, 1 + 1))
})
tar_make()
tar_script({
  tar_options()
  tar_pipeline(
    tar_target(y1, 1 + 1),
    tar_target(y2, 1 + 1),
    tar_target(z, y1 + y2)
  )
})
tar_make(starts_with("y")) # Only builds y1 and y2.
})
}
