% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_make_interactive.R
\name{tar_make_interactive}
\alias{tar_make_interactive}
\title{Interactive mode pipeline}
\usage{
tar_make_interactive(code)
}
\arguments{
\item{code}{Character vector of lines of a \verb{_targets.R} file
to define a pipeline.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Not a user-side function. Do not invoke directly.
Only exported to on a technicality.
}
\examples{
if (identical(Sys.getenv("TAR_INTERACTIVE_EXAMPLES"), "true")) {
tar_make_interactive("library(targets); tar_target(x, 123)")
message(x)
}
}
\keyword{internal}
