% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_make.R
\name{tar_make_reporters}
\alias{tar_make_reporters}
\title{Internal infrastructure function.}
\usage{
tar_make_reporters()
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
