% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_name.R
\name{tar_name}
\alias{tar_name}
\title{Get the name of the target currently running.}
\usage{
tar_name()
}
\description{
Get the name of the target currently running.
}
\details{
Must be invoked somewhere in the target's command.
Only works inside a \code{targets} pipeline.
}
\examples{
tar_target(returns_name, tar_name())
}
