% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_newer.R
\name{tar_newer}
\alias{tar_newer}
\title{List new targets}
\usage{
tar_newer(
  time,
  names = NULL,
  inclusive = FALSE,
  store = targets::tar_config_get("store")
)
}
\arguments{
\item{time}{A \code{POSIXct} object of length 1, time threshold.
Targets newer than this time stamp are returned.
For example, if \code{time = Sys.time - as.difftime(1, units = "weeks")}
then \code{tar_newer()} returns targets newer than one week ago.}

\item{names}{Names of eligible targets. Targets excluded from \code{names}
will not be returned even if they are newer than the given \code{time}.
You can supply symbols
or \code{tidyselect} helpers like \code{\link[=any_of]{any_of()}} and \code{\link[=starts_with]{starts_with()}}.
If \code{NULL}, all names are eligible.}

\item{inclusive}{Logical of length 1, whether to include targets
built at exactly the \code{time} given.}

\item{store}{Character of length 1, path to the
\code{targets} data store. Defaults to \code{tar_config_get("store")},
which in turn defaults to \verb{_targets/}.
When you set this argument, the value of \code{tar_config_get("store")}
is temporarily changed for the current function call.
See \code{\link[=tar_config_get]{tar_config_get()}} and \code{\link[=tar_config_set]{tar_config_set()}} for details
about how to set the data store path persistently
for a project.}
}
\value{
A character vector of names of old targets with recorded
timestamp metadata.
}
\description{
List all the targets whose last successful run occurred
after a certain point in time.
}
\details{
Only applies to targets with recorded time stamps:
just non-branching targets and individual dynamic branches.
As of \code{targets} version 0.6.0, these time
stamps are available for these targets regardless of
storage format. Earlier versions of \code{targets} do not record
time stamps for remote storage such as \code{format = "url"}
or \code{repository = "aws"} in \code{\link[=tar_target]{tar_target()}}.
}
\examples{
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temporary directory.
tar_script({
  list(tar_target(x, seq_len(2)))
}, ask = FALSE)
tar_make()
# targets newer than 1 week ago
tar_newer(Sys.time() - as.difftime(1, units = "weeks"))
# targets newer than 1 week from now
tar_newer(Sys.time() + as.difftime(1, units = "weeks"))
# Everything is still up to date.
tar_make()
# Invalidate all targets targets newer than 1 week ago
# so they run on the next tar_make().
invalidate_these <- tar_newer(Sys.time() - as.difftime(1, units = "weeks"))
tar_invalidate(any_of(invalidate_these))
tar_make()
})
}
}
\seealso{
Other time: 
\code{\link{tar_older}()},
\code{\link{tar_timestamp_raw}()},
\code{\link{tar_timestamp}()}
}
\concept{time}
