% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_noninteractive.R
\name{tar_noninteractive}
\alias{tar_noninteractive}
\title{Run if Target Markdown interactive mode is not on.}
\usage{
tar_noninteractive(code)
}
\arguments{
\item{code}{R code to run if Target Markdown interactive mode
is not turned on.}
}
\value{
If Target Markdown interactive mode is not turned on,
the function returns the result of running the code.
Otherwise, the function invisibly returns \code{NULL}.
}
\description{
In Target Markdown, run the enclosed code
only if interactive mode is not activated. Otherwise,
do not run the code.
}
\details{
Visit <books.ropensci.org/targets/markdown.html>
to learn about Target Markdown and interactive mode.
}
\examples{
tar_noninteractive(list(x = 1 + 1))
}
\seealso{
Other utilities: 
\code{\link{tar_cancel}()},
\code{\link{tar_envir}()},
\code{\link{tar_group}()},
\code{\link{tar_interactive}()},
\code{\link{tar_name}()},
\code{\link{tar_path}()},
\code{\link{tar_seed}()}
}
\concept{utilities}
