% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_objects.R
\name{tar_objects}
\alias{tar_objects}
\title{List saved targets}
\usage{
tar_objects()
}
\description{
List targets currently saved to \verb{_targets/objects/}.
Does not include dynamic files or cloud storage.
}
\examples{
if (FALSE) {
tar_dir({
tar_script({
  tar_option_set(workspace = "x")
  tar_pipeline(tar_target(x, "value"))
})
tar_make()
tar_workspaces()
})
}
}
