% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_option.R
\name{tar_option}
\alias{tar_option}
\title{Get a target option.}
\usage{
tar_option(option, default = NULL)
}
\arguments{
\item{option}{Character of length 1, name of an option to get.
Must be one of the argument names of \code{\link[=tar_options]{tar_options()}}.}

\item{default}{Default value if the option is unset (\code{NULL})}
}
\value{
Value of a target option. Returns \code{NULL} if the option is not set.
}
\description{
Get a target option. These options include default arguments to
\code{\link[=tar_target]{tar_target()}} such as packages, storage format,
iteration type, and cue.
Needs to be called before any calls to \code{\link[=tar_target]{tar_target()}}
in order to take effect.
}
\details{
This function goes well with \code{\link[=tar_target_raw]{tar_target_raw()}} when it comes
to defining external interfaces on top of the \code{targets} package to create
pipelines.
}
\examples{
\dontrun{
# Set target options:
tar_options(format = "fst_tbl")
tar_option("format")
# Reset target options:
tar_options()
tar_option("format")
}
}
