% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_options.R
\name{tar_option_export}
\alias{tar_option_export}
\title{Export options.}
\usage{
tar_option_export()
}
\value{
A list of options from tar_option_set().
}
\description{
Internal function. Not for users.
}
\keyword{internal}
