% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_option_reset.R
\name{tar_option_reset}
\alias{tar_option_reset}
\title{Reset target options.}
\usage{
tar_option_reset()
}
\value{
Nothing.
}
\description{
Reset target options you set with \code{\link[=tar_options]{tar_options()}}.
These options are mostly configurable default arguments to
\code{\link[=tar_target]{tar_target()}} and \code{\link[=tar_target_raw]{tar_target_raw()}}.
}
\examples{
tar_option("format") # default format before we set anything
tar_target(x, 1)$settings$format
tar_options(format = "fst_tbl") # new default format
tar_option("format")
tar_target(x, 1)$settings$format
tar_option_reset() # reset the format
tar_target(x, 1)$settings$format
}
