% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{tar_path_objects_dir}
\alias{tar_path_objects_dir}
\title{Path to directory of saved targets}
\usage{
tar_path_objects_dir(path_store)
}
\arguments{
\item{path_store}{Path to the data store.}

\item{name}{Target names.}
}
\description{
Internal function. Not for users.
}
\examples{
tar_path_objects_dir("_targets")
}
\keyword{internal}
