% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_path_script_support.R
\name{tar_path_script_support}
\alias{tar_path_script_support}
\title{Directory path to the support scripts
of the current target script}
\usage{
tar_path_script_support()
}
\value{
Character, directory path to the target script
of the pipeline currently running.
If called outside of the pipeline currently running,
\code{tar_path_script()} returns \code{tar_config_get("script")}.
}
\description{
Identify the directory path to the
support scripts
of the current target script
of the pipeline currently running.
}
\details{
A target script (default: \verb{_targets.R}) comes with
support scripts if it is written by Target Markdown.
These support scripts usually live in a folder called \verb{_targets_r/},
but the path may vary from case to case. The
\code{tar_path_scipt_support()} returns the path to
the folder with the support scripts.
}
\examples{
tar_path_script_support()
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temporary directory.
script <- tempfile()
tar_script(
  tar_target(x, tar_path_script_support()),
  script = script,
  ask = FALSE
)
tar_make(script = script)
tar_read(x)
})
}
}
\seealso{
Other utilities: 
\code{\link{tar_active}()},
\code{\link{tar_call}()},
\code{\link{tar_cancel}()},
\code{\link{tar_definition}()},
\code{\link{tar_envir}()},
\code{\link{tar_group}()},
\code{\link{tar_name}()},
\code{\link{tar_path_script}()},
\code{\link{tar_path_store}()},
\code{\link{tar_path_target}()},
\code{\link{tar_path}()},
\code{\link{tar_seed}()},
\code{\link{tar_source}()},
\code{\link{tar_store}()}
}
\concept{utilities}
