% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_prune.R
\name{tar_prune}
\alias{tar_prune}
\title{Remove targets that are no longer part of the pipeline.}
\usage{
tar_prune(callr_function = callr::r, callr_arguments = list())
}
\arguments{
\item{callr_function}{A function from \code{callr} to start a fresh clean R
process to do the work. Set to \code{NULL} to do the work in the current R
session, which is not reproducible and only recommended for
debugging purposes.}

\item{callr_arguments}{A list of arguments to \code{callr_function}.}
}
\value{
Nothing.
}
\description{
Remove target values from \verb{_targets/objects/} and
target metadata from \verb{_targets/meta/meta} for targets that are no longer
part of the pipeline. Global objects and dynamic files outside the
data store are unaffected. Also removes \verb{_targets/scratch/},
which is only needed while \code{\link[=tar_make]{tar_make()}}, \code{\link[=tar_make_clustermq]{tar_make_clustermq()}},
or \code{\link[=tar_make_future]{tar_make_future()}} is running.
}
\examples{
\dontrun{
tar_dir({
tar_script(
  tar_pipeline(
    tar_target(y1, 1 + 1),
    tar_target(y2, 1 + 1),
    tar_target(z, y1 + y2)
  )
)
tar_make()
# Remove some targets from the pipeline.
tar_script(tar_pipeline(tar_target(y1, 1 + 1)))
# Keep only the remaining targets in the data store.
tar_prune()
})
}
}
