% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{tar_random_name}
\alias{tar_random_name}
\title{Generate a random name}
\usage{
tar_random_name()
}
\value{
Character of length 1 with a random name.
}
\description{
Uses the lists of adjectives and animals from
\url{https://github.com/a-type/adjective-adjective-animal}.
by Grant Forrest under the MIT license.
}
\details{
Not a user-side function. Do not invoke directly.
}
\examples{
tar_random_name()
}
\keyword{internal}
