% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_random_port.R, R/utils_url.R
\name{tar_random_port}
\alias{tar_random_port}
\title{Random TCP port}
\usage{
tar_random_port(lower = 49152L, upper = 65355L)

tar_random_port(lower = 49152L, upper = 65355L)
}
\arguments{
\item{lower}{Integer of length 1, lowest possible port.}

\item{upper}{Integer of length 1, highest possible port.}
}
\value{
A random port not likely to be used by another process.

A random port not likely to be used by another process.
}
\description{
Not a user-side function. Exported for infrastructure
purposes only.

Not a user-side function. Exported for infrastructure
purposes only.
}
\examples{
if (requireNamespace("parallelly", quietly = TRUE)) {
tar_random_port()
}
if (requireNamespace("parallelly", quietly = TRUE)) {
tar_random_port()
}
}
\keyword{internal}
