% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_repository_cas_local.R
\name{tar_repository_cas_local_exists}
\alias{tar_repository_cas_local_exists}
\title{Existence check in local CAS.}
\usage{
tar_repository_cas_local_exists(cas, key)
}
\arguments{
\item{cas}{File path to the CAS repository.}

\item{key}{Key of the object in the CAS system.}
}
\value{
\code{TRUE} if the key exists in the CAS system, \code{FALSE} otherwise.
}
\description{
For internal use only.
}
\details{
\code{\link[=tar_repository_cas_local_exists]{tar_repository_cas_local_exists()}} uses an in-memory cache
in a package internal environment to maintain a list of keys that
exists. This avoids expensive one-time lookups to the file system
during \code{\link[=tar_make]{tar_make()}}.
}
\keyword{internal}
