% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_repository_cas_local.R
\name{tar_repository_cas_local_upload}
\alias{tar_repository_cas_local_upload}
\title{Local CAS upload.}
\usage{
tar_repository_cas_local_upload(cas, key, path)
}
\arguments{
\item{cas}{File path to the CAS repository.}

\item{key}{Key of the object in the CAS system.}

\item{path}{Staging path of the file.}
}
\value{
Called for its side effects.
}
\description{
For internal use only.
}
\keyword{internal}
