% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_resources_clustermq.R
\name{tar_resources_clustermq}
\alias{tar_resources_clustermq}
\title{Target resources: \code{clustermq}}
\usage{
tar_resources_clustermq(log_worker = FALSE, template = list())
}
\arguments{
\item{log_worker}{Logical of length 1, \code{log_worker}
argument to \code{clustermq::workers()}.}

\item{template}{Named list, \code{template} argument to
\code{clustermq::workers()}.}
}
\value{
Object of class \code{"tar_resources_clustermq"}, to be supplied
to the \code{clustermq} argument of \code{\link[=tar_resources]{tar_resources()}}.
}
\description{
Create the \code{clustermq} argument of \code{\link[=tar_resources]{tar_resources()}}]
to specify optional high-performance computing settings
for \code{\link[=tar_make_clustermq]{tar_make_clustermq()}}.
For details, see the documentation of the \code{clustermq} R package
and the corresponding argument names in this help file.
}
\section{Resources}{

Functions \code{\link[=tar_target]{tar_target()}} and \code{\link[=tar_option_set]{tar_option_set()}}
each takes an optional \code{resources} argument to supply
non-default settings of various optional backends for data storage
and high-performance computing. The \code{tar_resources()} function
is a helper to supply those settings in the correct manner.
}

\examples{
# Somewhere in you target script file (usually _targets.R):
tar_target(
  name,
  command(),
  resources = tar_resources(
    clustermq = tar_resources_clustermq(template = list(n_cores = 2))
  )
)
}
\seealso{
Other resources: 
\code{\link{tar_resources_aws}()},
\code{\link{tar_resources}()}
}
\concept{resources}
