% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_resources_url.R
\name{tar_resources_url}
\alias{tar_resources_url}
\title{Target resources: URL storage formats}
\usage{
tar_resources_url(handle = NULL)
}
\arguments{
\item{handle}{Object returned by \code{curl::new_handle} or \code{NULL}.}
}
\value{
Object of class \code{"tar_resources_url"}, to be supplied
to the url argument of \code{\link[=tar_resources]{tar_resources()}}.
}
\description{
Create the \code{url} argument of \code{\link[=tar_resources]{tar_resources()}}]
to specify optional settings for URL storage formats.
See the \code{format} argument of \code{\link[=tar_target]{tar_target()}} for details.
}
\section{Resources}{

Functions \code{\link[=tar_target]{tar_target()}} and \code{\link[=tar_option_set]{tar_option_set()}}
each takes an optional \code{resources} argument to supply
non-default settings of various optional backends for data storage
and high-performance computing. The \code{tar_resources()} function
is a helper to supply those settings in the correct manner.
}

\examples{
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
# Somewhere in you target script file (usually _targets.R):
tar_target(
  name,
  command(),
  format = "url",
  resources = tar_resources(
    url = tar_resources_url(handle = curl::new_handle())
  )
)
}
}
\seealso{
Other resources: 
\code{\link{tar_resources_aws}()},
\code{\link{tar_resources_clustermq}()},
\code{\link{tar_resources_feather}()},
\code{\link{tar_resources_fst}()},
\code{\link{tar_resources_future}()},
\code{\link{tar_resources_parquet}()},
\code{\link{tar_resources_qs}()},
\code{\link{tar_resources}()}
}
\concept{resources}
