% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_runtime.R
\name{tar_runtime_object}
\alias{tar_runtime_object}
\title{Get the \code{tar_runtime} object.}
\usage{
tar_runtime_object()
}
\value{
The internal \code{tar_runtime} object of class \code{"tar_runtime"}.
}
\description{
For internal purposes only. Not a user-side function.
Do not invoke directly.
}
\details{
Manages internal settings
that targets need while they run.
}
\examples{
tar_runtime_object()
}
\keyword{internal}
