% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_seed_set.R
\name{tar_seed_set}
\alias{tar_seed_set}
\title{Set a seed to run a target.}
\usage{
tar_seed_set(seed)
}
\arguments{
\item{seed}{Integer of length 1, value of the seed to set
with \code{set.seed()}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
\code{targets} generates its own target-specific seeds
using \code{\link[=tar_seed_create]{tar_seed_create()}}. Use \code{\link[=tar_seed_set]{tar_seed_set()}} to set one of
these seeds in R.
}
\details{
\code{\link[=tar_seed_set]{tar_seed_set()}} gives the user-supplied \code{seed} to
\code{set.seed()} and sets arguments \code{kind = "default"},
\code{normal.kind = "default"}, and \code{sample.kind = "default"}.
}
\examples{
seed <- tar_seed_create()
seed
sample(10)
tar_seed_set(seed)
sample(10)
tar_seed_set(seed)
sample(10)
}
\concept{seeds}
