% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stage.R
\name{tar_stage}
\alias{tar_stage}
\title{Identify the file path where a target will be staged.}
\usage{
tar_stage(default = NA_character_)
}
\arguments{
\item{default}{Character, value to return if \code{tar_stage()}
is called on its own outside a \code{targets} pipeline.
Having a default lets users run things without \code{\link[=tar_make]{tar_make()}},
which helps peel back layers of code and troubleshoot bugs.}
}
\value{
Character, staging file path.
}
\description{
Identify the file path where a target will be staged
just before storage. For reliability reasons,
\code{targets} writes to this staging file first and then moves
the staging file to the permanent path in the data store
(or the AWS S3 bucket in the case of AWS-backed storage formats).
}
\details{
The staging path is not available for dynamic files,
e.g. \code{format = "file"} or \code{"aws_file"}. These formats do not use
a special staging path at all.
}
\examples{
tar_stage()
tar_stage(tempfile())
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temporary directory.
tar_script(tar_target(returns_stage, tar_stage()), ask = FALSE)
tar_make()
tar_read(returns_stage)
})
}
}
\seealso{
Other utilities: 
\code{\link{tar_cancel}()},
\code{\link{tar_definition}()},
\code{\link{tar_envir}()},
\code{\link{tar_group}()},
\code{\link{tar_name}()},
\code{\link{tar_path}()},
\code{\link{tar_seed}()}
}
\concept{utilities}
