% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_store.R
\name{tar_store}
\alias{tar_store}
\title{Deprecated: current data store path}
\usage{
tar_store()
}
\value{
Character, file path to the data store
of the pipeline currently running.
If called outside of the pipeline currently running,
\code{tar_store()} returns \code{tar_config_get("store")}.
}
\description{
Deprecated: identify the file path to the data store
of the pipeline currently running.
}
\details{
\code{tar_store()} was deprecated on 2022-10-11 (version 0.13.5.9000).
Use \code{\link[=tar_path_store]{tar_path_store()}} instead.
}
\examples{
tar_path_store()
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script(tar_target(x, tar_path_store()), ask = FALSE)
store <- tempfile()
tar_make(store = store)
tar_read(x, store = store)
})
}
}
\seealso{
Other utilities: 
\code{\link{tar_active}()},
\code{\link{tar_call}()},
\code{\link{tar_cancel}()},
\code{\link{tar_definition}()},
\code{\link{tar_envir}()},
\code{\link{tar_group}()},
\code{\link{tar_name}()},
\code{\link{tar_path_script_support}()},
\code{\link{tar_path_script}()},
\code{\link{tar_path_store}()},
\code{\link{tar_path_target}()},
\code{\link{tar_path}()},
\code{\link{tar_seed}()},
\code{\link{tar_source}()}
}
\concept{utilities}
\keyword{internal}
