% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_tidyselect.R
\name{tar_tidyselect}
\alias{tar_tidyselect}
\title{Internal infrastructure function.}
\usage{
tar_tidyselect(names_quosure, choices)
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
