% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_toggle.R
\name{tar_toggle}
\alias{tar_toggle}
\title{Choose code to run based on Target Markdown mode.}
\usage{
tar_toggle(interactive, noninteractive)
}
\arguments{
\item{interactive}{R code to run if Target Markdown interactive mode is
activated.}

\item{noninteractive}{R code to run if Target Markdown interactive mode is
not activated.}
}
\value{
If Target Markdown interactive mode is not turned on,
the function returns the result of running the code.
Otherwise, the function invisibly returns \code{NULL}.
}
\description{
Run one piece of code if Target Markdown mode
interactive mode is turned on and another piece of code otherwise.
}
\details{
Visit <books.ropensci.org/targets/markdown.html>
to learn about Target Markdown and interactive mode.
}
\examples{
tar_toggle(
  message("In interactive mode."),
  message("Not in interactive mode.")
)
}
\seealso{
Other Target Markdown: 
\code{\link{tar_engine_knitr}()},
\code{\link{tar_interactive}()},
\code{\link{tar_noninteractive}()}
}
\concept{Target Markdown}
