% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_traceback.R
\name{tar_traceback}
\alias{tar_traceback}
\title{Get the traceback of the last errored target.}
\usage{
tar_traceback()
}
\description{
Get the traceback of the last errored target.
Useful for debugging purposes. Requires
\code{callr_function = NULL} in \code{\link[=tar_make]{tar_make()}}.
}
\examples{
\dontrun{
tar_dir({
tar_script(
   tar_pipelinetar_target(x, sqrt(sqrt(stop("error message"))))
)
# tar_make(callr_function = NULL)
# tar_traceback()
})
}
}
