% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_unscript.R
\name{tar_unscript}
\alias{tar_unscript}
\title{Remove files from \verb{_targets_r/}.}
\usage{
tar_unscript()
}
\value{
\code{NULL} (invisibly).
}
\description{
Target Markdown writes files to \verb{_targets_r/}
at the project root (next to \verb{_targets.R}).
Use \code{\link[=tar_unscript]{tar_unscript()}} to remove all of \verb{_targets_r/} and start with
a fresh set of targets and globals to define the pipeline.
}
\examples{
tar_dir({ # tar_dir() runs code from a temporary directory.
tar_unscript()
})
}
