% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_watch.R
\name{tar_watch_port}
\alias{tar_watch_port}
\title{Random port for \code{\link[=tar_watch]{tar_watch()}}}
\usage{
tar_watch_port(lower = 49152L, upper = 65355L)
}
\arguments{
\item{lower}{Integer of length 1, lowest possible port.}

\item{upper}{Integer of length 1, highest possible port.}
}
\value{
A random port not likely to be used by another process.
}
\description{
Required for \code{\link[=tar_watch]{tar_watch()}}. Not a user-side function.
}
\examples{
tar_watch_port()
}
\keyword{internal}
