% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_watch_server.R
\name{tar_watch_server}
\alias{tar_watch_server}
\title{Shiny module server for tar_watch()}
\usage{
tar_watch_server(
  id,
  height = "650px",
  exclude = ".Random.seed",
  config = Sys.getenv("TAR_CONFIG", "_targets.yaml")
)
}
\arguments{
\item{id}{Character of length 1, ID corresponding to the UI function
of the module.}

\item{height}{Character of length 1,
height of the \code{visNetwork} widget and branches table.}

\item{exclude}{Character vector of nodes to omit from the graph.}

\item{config}{Character of length 1, file path of the YAML
configuration file with \code{targets} project settings.
The \code{config} argument specifies which YAML configuration
file that \code{tar_config_get()} reads from or \code{tar_config_set()}
writes to in a single function call.
It does not globally change which configuration file is used
in subsequent function calls. The default file path of the YAML
file is always \verb{_targets.yaml} unless you set another
default path using the \code{TAR_CONFIG} environment variable,
e.g. \verb{Sys.setenv(TAR_CONFIG = "custom.yaml)}.
But please keep in mind that unless you set
\code{TAR_CONFIG} permanently inside your \code{.Renviron} file, the
environment variable will reset when your R session restarts.
Please check \code{Sys.getenv("TAR_CONFIG")} to be sure.}
}
\value{
A Shiny module server.
}
\description{
Use \code{\link[=tar_watch_ui]{tar_watch_ui()}} and \code{tar_watch_server()}
to include \code{\link[=tar_watch]{tar_watch()}} as a Shiny module in an app.
}
\seealso{
Other progress: 
\code{\link{tar_built}()},
\code{\link{tar_canceled}()},
\code{\link{tar_errored}()},
\code{\link{tar_poll}()},
\code{\link{tar_progress_branches}()},
\code{\link{tar_progress_summary}()},
\code{\link{tar_progress}()},
\code{\link{tar_skipped}()},
\code{\link{tar_started}()},
\code{\link{tar_watch_ui}()},
\code{\link{tar_watch}()}
}
\concept{progress}
