% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_workspace.R
\name{tar_workspace}
\alias{tar_workspace}
\title{Load a saved workspace and seed for debugging.}
\usage{
tar_workspace(name, envir = parent.frame(), packages = TRUE, source = TRUE)
}
\arguments{
\item{name}{Symbol, name of the target whose workspace to read.}

\item{envir}{Environment in which to put the objects.}

\item{packages}{Logical, whether to load the required packages
of the target.}

\item{source}{Logical, whether to run \verb{_targets.R} to load user-defined
global object dependencies into \code{envir}. If \code{TRUE}, then \code{envir}
should either be the global environment or inherit from the
global environment.}
}
\value{
the function returns no value, but it does load
the target's required packages, as well as multiple objects
into the environment (\code{envir} argument) in order to replicate the
workspace where the error happened. These objects include
the global objects at the time \code{\link[=tar_make]{tar_make()}} was called and the
dependency targets. The random number generator seed for the
target is also assigned with \code{set.seed()}
}
\description{
Load the packages, workspace, and random number generator seed
of an errored target attempted with \code{error = "workspace"}.
Remove workspace files with \code{\link[=tar_undebug]{tar_undebug()}} when you are done debugging.
}
\details{
If you set \code{error = "workspace"} in \code{\link[=tar_option_set]{tar_option_set()}}
or \code{\link[=tar_target]{tar_target()}}, then if that target throws an error
in \code{\link[=tar_make]{tar_make()}}, it will save its workspace to an RDS file
in \verb{_targets/workspaces/}. The workspace is a compact reference
that allows \code{tar_workspace()} to load the target's dependencies
and random number generator seed as long as the data objects
are still in the data store (usually files in \verb{_targets/objects/}).
}
\examples{
\dontrun{
tmp <- sample(1)
tar_script({
  tar_option_set(error = "workspace") # Required for saving workspaces.
  tar_pipeline(
    tar_target(x, "loaded"),
    tar_target(y, stop(x))
  )
})
try(tar_make())
exists("x") # Should be FALSE.
tail(.Random.seed)
tar_workspace(y)
exists("x") # Should be TRUE.
print(x) # "loaded"
tail(.Random.seed) # Should be different.
}
}
