% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_workspace_policy.R
\name{tar_workspace_policy}
\alias{tar_workspace_policy}
\title{Workspace policy}
\usage{
tar_workspace_policy(
  always = character(0),
  never = character(0),
  error = FALSE
)
}
\arguments{
\item{always}{Character vector of names of targets to always save
workspaces. Must not include any elements of \code{never}.}

\item{never}{Character vector of names of targets to never save
workspaces. Must not include any elements of \code{always}.}

\item{error}{Logical, whether to save a workspace for each
target that throws an error (except for the targets in \code{never}).}
}
\value{
An object of class \code{"tar_worspace_policy"}. Supply this object
to the \code{workspaces} argument of \code{\link[=tar_option_set]{tar_option_set()}} to control
when \code{targets} saves workspaces.
}
\description{
Control when \code{targets} saves workspace files.
}
\details{
See \code{\link[=tar_workspace]{tar_workspace()}} for details
}
\examples{
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temporary directory.
tar_script({
  tar_option_set(workspaces = tar_workspace_policy(error = TRUE))
  list(
    tar_target(x, "value"),
    tar_target(y, x)
  )
}, ask = FALSE)
tar_make()
tar_workspaces()
tar_workspaces(contains("x"))
})
}
}
\seealso{
Other debug: 
\code{\link{tar_load_globals}()},
\code{\link{tar_traceback}()},
\code{\link{tar_workspaces}()},
\code{\link{tar_workspace}()}
}
\concept{debug}
