% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_workspaces.R
\name{tar_workspaces}
\alias{tar_workspaces}
\title{List saved target workspaces.}
\usage{
tar_workspaces()
}
\description{
List target workspaces currently saved to
\verb{_targets/workspaces/}. See \code{\link[=tar_workspace]{tar_workspace()}} for more information.
}
\examples{
if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
tar_dir({
tar_script({
  tar_option_set(error = "workspace")
  tar_pipeline(
    tar_target(x, "value"),
    tar_target(y, x)
  )
})
tar_make()
tar_workspaces()
})
}
}
