% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_workspaces_destroy.R
\name{tar_workspaces_destroy}
\alias{tar_workspaces_destroy}
\title{Remove debugging files.}
\usage{
tar_workspaces_destroy()
}
\value{
Nothing.
}
\description{
Destroy \verb{_targets/workspaces/} and \verb{_targets/scratch/},
which are directories with files for debugging.
}
\examples{
\dontrun{
tar_script({
  tar_option_set(error = "workspace") # Required for saving workspaces.
  tar_pipeline(
    tar_target(x, "loaded"),
    tar_target(y, stop(x))
  )
})
try(tar_make())
file.exists("_targets/workspaces/y") # TRUE
tar_workspaces_destroy()
file.exists("_targets/workspaces/y") # FALSE
}
}
