% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_conclude}
\alias{target_conclude}
\title{Internal infrastructure function.}
\usage{
target_conclude(target, pipeline, scheduler, meta)
}
\arguments{
\item{target}{Target object.}

\item{pipeline}{Pipeline object.}

\item{scheduler}{Scheduler object.}

\item{meta}{Meta object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
