% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_prepare}
\alias{target_prepare}
\title{Internal infrastructure function.}
\usage{
target_prepare(target, pipeline, scheduler)
}
\arguments{
\item{target}{Target object.}

\item{pipeline}{Pipeline object.}

\item{scheduler}{Scheduler object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
