% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_read_value}
\alias{target_read_value}
\title{Internal infrastructure function.}
\usage{
target_read_value(target, pipeline)
}
\arguments{
\item{target}{Target object.}

\item{pipeline}{Pipeline object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
