% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_run}
\alias{target_run}
\title{Internal infrastructure function.}
\usage{
target_run(target)
}
\arguments{
\item{target}{Target object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
