% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_run_remote}
\alias{target_run_remote}
\title{Internal function to run a target on a remote worker.}
\usage{
target_run_remote(target, garbage_collection)
}
\arguments{
\item{target}{A target object.}

\item{garbage_collection}{Logical, whether to run garbage collection.}
}
\description{
For internal purposes only. Not a user-side function.
}
\keyword{internal}
