% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_run_worker}
\alias{target_run_worker}
\title{Internal function to run a target on a worker.}
\usage{
target_run_worker(target, envir, config)
}
\arguments{
\item{target}{A target object.}

\item{envir}{An environment.}

\item{config}{List, and exported object of class \code{"tar_config"}.}
}
\description{
For internal purposes only. Not a user-side function.
}
\keyword{internal}
