% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_should_run_remote}
\alias{target_should_run_remote}
\title{Internal infrastructure function.}
\usage{
target_should_run_remote(target)
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
