% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_update_depend}
\alias{target_update_depend}
\title{Internal infrastructure function.}
\usage{
target_update_depend(target, meta)
}
\arguments{
\item{target}{Target object.}

\item{meta}{Meta object.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
