% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_package.R
\docType{package}
\name{targets-package}
\alias{targets-package}
\title{targets: Dynamic Function-Oriented Make-Like Declarative Pipelines for R}
\description{
The targets package is a pipeline toolkit that brings together
function-oriented programming and Make-like declarative pipelines for
Statistics and data science in R. It implements a pipeline as collection
of interconnected tasks, analyzes the dependency relationships among these
tasks, skips steps that are already up to date, runs the necessary
computations with optional parallel workers, abstracts files as
R objects, and provides tangible evidence that the results match
the underlying code and data. The methodology in this package
borrows from GNU Make by Richard Stallman (2015, ISBN:978-9881443519)
and drake by Will Landau (2018) (doi:10.21105/joss.00550).
}
