% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{targets_manual_empty}
\alias{targets_manual_empty}
\title{Empty function}
\usage{
targets_manual_empty()
}
\value{
Nothing.
}
\description{
This is a bookdown manual. It only
borrows R package infrastructure for devops purposes.
}
\examples{
targets_manual_empty()
}
