% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{trn}
\alias{trn}
\title{Internal infrastructure function.}
\usage{
trn(condition, x, y)
}
\arguments{
\item{condition}{Logical, whether to choose x or y.}

\item{x}{Object to choose of \code{condition} is \code{TRUE}.}

\item{y}{Object to choose of \code{condition} is \code{FALSE}.}
}
\description{
Not a user-side function.
Only use for developing external HPC backend packages.
}
\keyword{internal}
