% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_targets.R
\name{use_targets_scheduler}
\alias{use_targets_scheduler}
\title{Detect the scheduler on the system.}
\usage{
use_targets_scheduler()
}
\value{
Character of length 1 with the name of the resource
manager. One of the following:
\itemize{
\item \code{"multicore"}: local forked processes on Linux-like systems.
\item \code{"multiprocess"}: local platform-independent and multi-process.
\item \code{"slurm"}: SLURM clusters.
\item \code{"sge"}: Sun Grid Engine clusters.
\item \code{"lsf"}: LSF clusters.
}
}
\description{
Automatically detect the resource manager
on the system.
}
\details{
Not a user-side function. Do not invoke directly.
}
\examples{
use_targets_scheduler()
}
\keyword{internal}
