% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pipeline.R
\name{pipeline_validate_lite}
\alias{pipeline_validate_lite}
\title{Abridged pipeline validation function.}
\usage{
pipeline_validate_lite(pipeline)
}
\arguments{
\item{pipeline}{A pipeline object.}
}
\description{
Internal function. Do not invoke directly.
}
\keyword{internal}
