% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_bind.R
\name{tar_bind}
\alias{tar_bind}
\title{Combine pipeline objects (deprecated).}
\usage{
tar_bind(...)
}
\arguments{
\item{...}{Pipeline objects or nested lists of pipeline objects.
You can generate a pipeline object with \code{\link[=tar_pipeline]{tar_pipeline()}}.}
}
\description{
Functions \code{tar_pipeline()} and \code{\link[=tar_bind]{tar_bind()}} are deprecated.
Instead, simply end your \verb{_targets.R} file with a list of target objects.
You can nest these objects however you like.
}
\details{
Deprecated on 2021-01-03.
}
\examples{
# In _targets.R:
library(targets)
list( # You no longer need tar_pipeline() here.
  tar_target(data_file, "data.csv", format = "file"),
  list( # Target lists can be arbitrarily nested.
    tar_target(data_object, read.csv(data_file)),
    tar_target(analysis, analyze(data_object))
  )
)
}
\keyword{internal}
