% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_dir.R
\name{tar_dir}
\alias{tar_dir}
\title{Execute code in a temporary directory.}
\usage{
tar_dir(code)
}
\arguments{
\item{code}{User-defined code.}
}
\value{
Return value of the user-defined code.
}
\description{
Runs code inside a \code{tempfile()} directory
using \code{withr::with_dir()}. Handy for writing examples in
\href{https://wlandau.github.io/targetopia.html}{targetopia} packages
(extensions to \code{targets} tailored to specific use cases).
}
\examples{
tar_dir(file.create("only_exists_in_tar_dir"))
file.exists("only_exists_in_tar_dir")
}
