% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_edit.R
\name{tar_edit}
\alias{tar_edit}
\title{Open _targets.R for editing.}
\usage{
tar_edit()
}
\description{
Looks for _targets.R in the current working directory.
Requires the \code{usethis} package.
}
\seealso{
Other scripts: 
\code{\link{tar_github_actions}()},
\code{\link{tar_helper_raw}()},
\code{\link{tar_helper}()},
\code{\link{tar_renv}()},
\code{\link{tar_script}()}
}
\concept{scripts}
