% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_exist_script.R
\name{tar_exist_script}
\alias{tar_exist_script}
\title{Check if the target script exists.}
\usage{
tar_exist_script()
}
\value{
Logical of length 1, whether the current project's metadata exists.
}
\description{
Check if the \verb{_targets.R} file of the current project exists.
exists for the current project.
}
\examples{
tar_exist_script()
}
\seealso{
Other existence: 
\code{\link{tar_exist_meta}()},
\code{\link{tar_exist_objects}()},
\code{\link{tar_exist_process}()},
\code{\link{tar_exist_progress}()}
}
\concept{existence}
