% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_mermaid.R
\name{tar_mermaid}
\alias{tar_mermaid}
\title{\code{mermaid.js} dependency graph.}
\usage{
tar_mermaid(
  targets_only = FALSE,
  names = NULL,
  shortcut = FALSE,
  allow = NULL,
  exclude = ".Random.seed",
  outdated = TRUE,
  label = NULL,
  reporter = targets::tar_config_get("reporter_outdated"),
  callr_function = callr::r,
  callr_arguments = targets::callr_args_default(callr_function),
  envir = parent.frame(),
  script = targets::tar_config_get("script"),
  store = targets::tar_config_get("store")
)
}
\value{
A character vector of lines of code of the \code{mermaid.js} graph.
You can visualize the graph by copying the text
into a public online \code{mermaid.js} editor or a \code{mermaid} GitHub code chunk
(\url{https://github.blog/2022-02-14-include-diagrams-markdown-files-mermaid}). # nolint
}
\description{
Visualize the dependency graph with a static \code{mermaid.js} graph.
}
\details{
\code{mermaid.js} is a JavaScript library for constructing
static visualizations of graphs.
}
\examples{
if (identical(Sys.getenv("TAR_INTERACTIVE_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temporary directory.
tar_script({
  tar_option_set()
  list(
    tar_target(y1, 1 + 1),
    tar_target(y2, 1 + 1),
    tar_target(z, y1 + y2)
  )
})
# Copy the text into a mermaid.js online editor
# or a mermaid GitHub code chunk:
tar_mermaid()
})
}
}
\seealso{
Other visualize: 
\code{\link{tar_glimpse}()},
\code{\link{tar_visnetwork}()}
}
\concept{visualize}
