% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_option.R
\name{tar_option}
\alias{tar_option}
\title{Get a target option.}
\usage{
tar_option(option)
}
\arguments{
\item{option}{Character of length 1, name of an option to get.
Must be one of the argument names of \code{\link[=tar_options]{tar_options()}}.}
}
\value{
Value of a target option.
}
\description{
Get a target option. These options include default arguments to
\code{\link[=tar_target]{tar_target()}} such as packages, storage format,
iteration type, and cue.
Needs to be called before any calls to \code{\link[=tar_target]{tar_target()}}
in order to take effect.
}
\details{
This function goes well with \code{\link[=tar_target_raw]{tar_target_raw()}} when it comes
to defining external interfaces on top of the \code{targets} package to create
pipelines.
}
\examples{
tar_option("format") # default format before we set anything
tar_target(x, 1)$settings$format
tar_options(format = "fst_tbl") # new default format
tar_option("format")
tar_target(x, 1)$settings$format
tar_options(format = "rds") # reset the format
}
