% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_path.R
\name{tar_path}
\alias{tar_path}
\title{Get the file path of a target's stored value.}
\usage{
tar_path(name = NULL)
}
\arguments{
\item{name}{Symbol, name of the target.
If \code{NULL}, \code{tar_path()} returns the path of the target currently running
in a pipeline.}
}
\value{
File path to the target.
}
\description{
Get the file path of a target's stored value. Could be the
target currently running in a pipeline or a target of your choice.
}
\details{
Call \code{tar_path(name = "your_target")} to get the file path
of a selected target. Inside the command of a target, call
\code{tar_path()} with no arguments to get the path to the target
currently running. This function does not apply to dynamic files
(targets with \code{format = "file"})
because those paths are not known in advance.
}
\examples{
tar_path(your_target)
tar_target(returns_path, tar_path())
}
