% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_pipeline.R
\name{tar_pipeline}
\alias{tar_pipeline}
\title{Declare a pipeline (deprecated).}
\usage{
tar_pipeline(...)
}
\arguments{
\item{...}{Targets or lists of targets defined with \code{\link[=tar_target]{tar_target()}}.}
}
\value{
A pipeline object.
}
\description{
Functions \code{tar_pipeline()} and \code{\link[=tar_bind]{tar_bind()}} are deprecated.
Instead, simply end your \verb{_targets.R} file with a list of target objects.
You can nest these objects however you like.
}
\details{
Deprecated on 2021-01-03.
}
\examples{
# In _targets.R:
library(targets)
list( # You no longer need tar_pipeline() here.
  tar_target(data_file, "data.csv", format = "file"),
  list( # Target lists can be arbitrarily nested.
    tar_target(data_object, read.csv(data_file)),
    tar_target(analysis, analyze(data_object))
  )
)
}
\keyword{internal}
