% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_repository_cas.R
\name{tar_repository_cas}
\alias{tar_repository_cas}
\title{Define a custom content-addressable storage (CAS) repository.}
\usage{
tar_repository_cas(upload, download, exists)
}
\arguments{
\item{upload}{A function with arguments \code{path} and \code{key}, in that order.
This function should upload the file or directory at \code{path}
to the CAS system. \code{key} denotes the name of the destination data object
in the CAS system.}

\item{download}{A function with arguments \code{path} and \code{key}, in that order.
This function should download the data object at \code{key} from
the CAS system to the file or directory at \code{path}.
to the CAS system. \code{key} denotes the name of the destination data object
in the CAS system.}

\item{exists}{A function with a single argument \code{key}.
This function should check if there is an object at \code{key} in
the CAS system.'}
}
\description{
Define a custom storage repository that uses
content-addressable storage (CAS).
}
\details{
Repository functions:
In \code{\link[=tar_repository_cas]{tar_repository_cas()}}, functions \code{upload}, \code{download},
and \code{exists} must be perfectly pure
and perfectly self-sufficient.
They must load or namespace all their own packages,
and they must not depend on any custom user-defined
functions or objects in the global environment of your pipeline.
\code{targets} converts each function to and from text,
so it must not rely on any data in the closure.
This disqualifies functions produced by \code{Vectorize()},
for example.

Some functions may need to be adapted and configured based on other
inputs. For example, you may want to define
\verb{upload = \\(path, key) file.move(path, file.path(folder, key))}
but do not want to hard-code a value of \code{folder} when you write the
underlying function. \code{substitute()} can help inject values as needed.
For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{upload = substitute(
  \\(path, key) file.move(path, file.path(folder, key)),
  list(folder = "my_cas")
)
}\if{html}{\out{</div>}}
}
\seealso{
Other targets: 
\code{\link{tar_cue}()},
\code{\link{tar_format}()},
\code{\link{tar_target}()},
\code{\link{tar_target_raw}()}
}
\concept{targets}
