% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_seed.R
\name{tar_seed}
\alias{tar_seed}
\title{Deprecated: get the seed of the current target.}
\usage{
tar_seed(default = 1L)
}
\arguments{
\item{default}{Integer, value to return if \code{tar_seed_get()}
is called on its own outside a \code{targets} pipeline.
Having a default lets users run things without \code{\link[=tar_make]{tar_make()}},
which helps peel back layers of code and troubleshoot bugs.}
}
\value{
Integer of length 1. If invoked inside a \code{targets} pipeline,
the return value is the seed of the target currently running,
which is a deterministic function of the target name. Otherwise,
the return value is \code{default}.
}
\description{
Deprecated on 2023-10-12 (\code{targets} version 1.3.2.9001).
Use \code{\link[=tar_seed_get]{tar_seed_get()}} instead.
}
\examples{
tar_seed_get()
tar_seed_get(default = 123L)
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script(tar_target(returns_seed, tar_seed_get()), ask = FALSE)
tar_make()
tar_read(returns_seed)
})
}
}
\keyword{internal}
