% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_timestamp.R
\name{tar_timestamp}
\alias{tar_timestamp}
\title{Get the timestamp of a target.}
\usage{
tar_timestamp(name = NULL)
}
\arguments{
\item{name}{Symbol, name of the target. If \code{NULL} (default)
then \code{tar_timestamp()} will attempt to return the timestamp
of the target currently running. Must be called inside a target's
command or a supporting function in order to work.}
}
\value{
A POSIXct object. If the target does not exists,
the return value is \code{as.POSIXct("0000-01-01")}.
If the target has a local storage
format, the return value is the maximum timestamp
over all the files. (Targets with \code{format = "file"}
may have multiple files.) For AWS-backed formats
such as \code{"aws_parquet"}, the time stamp is obtained
from the modification timestamp of the AWS S3 object header.
}
\description{
Get the time that a target's data was last modified.
}
\details{
If the files are stored locally, the time
}
\examples{
if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temporary directory.
tar_script({
  list(tar_target(x, 1))
}, ask = FALSE)
tar_make()
# Get the timestamp.
tar_timestamp(x)
# We can use the timestamp to cancel the target
# if it already ran within the last hour.
# Be sure to set `cue = tar_cue(mode = "always")`
# if you want the target to always check the timestamp.
tar_script({
  list(
  tar_target(
    x,
    tar_cancel((Sys.time() - tar_timestamp()) < 3600),
    cue = tar_cue(mode = "always")
  )
)}, ask = FALSE)
tar_make()
})
}
}
\seealso{
Other utilities: 
\code{\link{tar_cancel}()},
\code{\link{tar_envir}()},
\code{\link{tar_group}()},
\code{\link{tar_name}()},
\code{\link{tar_path}()},
\code{\link{tar_seed}()}
}
\concept{utilities}
