% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_unscript.R
\name{tar_unscript}
\alias{tar_unscript}
\title{Remove target script helper files.}
\usage{
tar_unscript()
}
\value{
\code{NULL} (invisibly).
}
\description{
Target Markdown writes helper scripts in a folder accompanying
the target script. If the target script is the default \verb{_targets.R},
the helper scripts are in \verb{_targets_r/}. If the some other location
like \code{custom/script.R}, the helper scripts are in \verb{custom/script_r/}.
Use \code{\link[=tar_unscript]{tar_unscript()}} to remove all these helper scripts.
}
\examples{
tar_dir({ # tar_dir() runs code from a temporary directory.
tar_unscript()
})
}
