% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_watch.R
\name{tar_watch}
\alias{tar_watch}
\title{Shiny app to watch the dependency graph.}
\usage{
tar_watch()
}
\description{
Calls \code{\link[=tar_visnetwork]{tar_visnetwork()}} and offers a GUI
to select settings.
}
\examples{
tar_dir({
tar_script()
# tar_watch()
})
}
