% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_workspace.R
\name{tar_workspace}
\alias{tar_workspace}
\title{Load a saved workspace and seed for debugging.}
\usage{
tar_workspace(name, envir = parent.frame())
}
\arguments{
\item{name}{Symbol, name of the target whose workspace to read.}

\item{envir}{Environment in which to put the objects.}
}
\value{
the function returns no value, but it does load multiple objects
into the environment (\code{envir} argument) in order to replicate the
workspace where the error happened. These objects include
the global objects at the time \code{\link[=tar_make]{tar_make()}} was called, the
dependency targets, the random number generator seed
(assigned to \code{.tar_seed} and set with \code{set.seed()})
and the traceback of the error (assigned to \code{.tar_traceback}).
}
\description{
Load the workspace and random number generator seed
of an errored target attempted with \code{error = "save"}.
}
\details{
If you set \code{error = "save"} in \code{\link[=tar_option_set]{tar_option_set()}}
or \code{\link[=tar_target]{tar_target()}}, then if that target throws an error
in \code{\link[=tar_make]{tar_make()}}, it will save its workspace to a compressed file
in \verb{_targets/workspaces/}. The workspace includes the
global objects at the time \code{\link[=tar_make]{tar_make()}} was called, the
dependency targets, the random number generator seed
(assigned to \code{.tar_seed} and set with \code{set.seed()})
and the traceback of the error (assigned to \code{.tar_traceback}).
}
\examples{
\dontrun{
tmp <- sample(1)
tar_script({
  tar_option_set(error = "save") # Required for saving workspaces.
  tar_pipeline(
    tar_target(x, "loaded"),
    tar_target(y, stop(x))
  )
})
try(tar_make())
exists("x") # Should be FALSE.
tail(.Random.seed)
tar_workspace(y)
exists("x") # Should be TRUE.
print(x) # "loaded"
tail(.Random.seed) # Should be different.
tail(.tar_traceback, 1)
}
}
