% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_run_worker}
\alias{target_run_worker}
\title{Internal function to run a target on a worker.}
\usage{
target_run_worker(target, envir, options, config)
}
\arguments{
\item{target}{A target object.}

\item{envir}{An environment or the string \code{"globalenv"}.}

\item{options}{List, exported from an object of class \code{"tar_options"}.}

\item{config}{List, exported from an object of class \code{"tar_config"}.}
}
\description{
For internal purposes only. Not a user-side function.
}
\keyword{internal}
