% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_target.R
\name{target_run_worker}
\alias{target_run_worker}
\title{Internal function to run a target on a worker.}
\usage{
target_run_worker(target)
}
\arguments{
\item{target}{A target object.}
}
\description{
For internal purposes only. Not a user-side function.
}
\keyword{internal}
