% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_package.R
\docType{package}
\name{targets-package}
\alias{targets-package}
\title{targets: Dynamic Function-Oriented Make-Like Declarative Pipelines for R}
\description{
A pipeline toolkit for Statistics and data science in R,
the \code{targets} package brings function-oriented programming to
Make-like declarative pipelines. \code{targets} orchestrates a pipeline
as a graph of dependencies,
skips steps that are already up to date, runs the necessary
computations with optional parallel workers, abstracts files as
R objects, and provides tangible evidence that the results are
reproducible given the underlying code and data.
The methodology in this package
borrows from GNU Make (2015, ISBN:978-9881443519)
and \code{drake} (2018, \doi{doi:10.21105/joss.00550}).
}
\seealso{
Other help: 
\code{\link{tar_reprex}()},
\code{\link{use_targets_rmd}()},
\code{\link{use_targets}()}
}
\concept{help}
