% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_targets.R
\name{use_targets}
\alias{use_targets}
\title{Use targets}
\usage{
use_targets(path = "_targets.Rmd", open = interactive())
}
\arguments{
\item{path}{Character of length 1, output path of the
Target Markdown report relative to the current active
project.}

\item{open}{Logical, whether to open the file for editing
in the RStudio IDE.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Create an example Target Markdown report
to get started with {targets}.
}
\examples{
if (identical(Sys.getenv("TAR_INTERACTIVE_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temporary directory.
use_targets(open = FALSE)
})
}
}
\seealso{
Other help: 
\code{\link{tar_reprex}()},
\code{\link{targets-package}}
}
\concept{help}
